Custom hook for all things chart
